이미지 타입 상수를 모듈로 분리하여 중복 제거/API 관심사 분리 및 함수명 통일
