+++ src/app/_apis/axios-client.ts
@@ -1,0 +1,111 @@
+import axios, { AxiosInstance, AxiosResponse } from 'axios';
+
+const baseURL = process.env.NEXT_PUBLIC_API_URL;
+
+// API 클라이언트 기본 설정
+const defaultHeaders = {
+  'Content-Type': 'application/json',
+};
+
+// API 요청 래퍼 함수
+const apiClient: AxiosInstance = axios.create({
+  baseURL,
+  headers: defaultHeaders,
+});
+
+apiClient.interceptors.request.use(
+  (config) => {
+    if (typeof window !== 'undefined') {
+      const token = localStorage.getItem('token');
+      if (token) {
+        config.headers.Authorization = `Bearer ${token}`;
+      }
+    }
+    if (config.data instanceof FormData) {
+      delete config.headers['Content-Type'];
+    }
+    return config;
+  },
+  (error) => {
+    return Promise.reject(error);
+  }
+);
+
+apiClient.interceptors.response.use(
+  (response) => {
+    return response;
+  },
+  (error) => {
+    if (axios.isAxiosError(error) && error.response?.status === 401) {
+      // Handle unauthorized access
+      if (typeof window !== 'undefined') {
+        localStorage.removeItem('token');
+        window.location.href = '/login';
+      }
+    }
+    return Promise.reject(error);
+  }
+);
+
+export const request = async <T>(
+  config: Parameters<typeof apiClient.request>[0]
+): Promise<T> => {
+  try {
+    const response: AxiosResponse<T> = await apiClient.request(config);
+    return response.data;
+  } catch (error) {
+    throw error;
+  }
+};
+
+export const get = async <T>(
+  endpoint: string,
+  params?: Record<string, string | number | boolean>
+): Promise<T> => {
+  const url = new URL(endpoint, baseURL);
+  if (params) {
+    Object.entries(params).forEach(([key, value]) => {
+      if (value !== undefined && value !== null) {
+        url.searchParams.append(key, String(value));
+      }
+    });
+  }
+  return request<T>({
+    method: 'GET',
+    url: url.toString(),
+  });
+};
+
+export const post = async <T>(endpoint: string, data?: unknown): Promise<T> => {
+  return request<T>({
+    method: 'POST',
+    url: endpoint,
+    data,
+  });
+};
+
+export const put = async <T>(endpoint: string, data?: unknown): Promise<T> => {
+  return request<T>({
+    method: 'PUT',
+    url: endpoint,
+    data,
+  });
+};
+
+export const patch = async <T>(
+  endpoint: string,
+  data?: unknown
+): Promise<T> => {
+  return request<T>({
+    method: 'PATCH',
+    url: endpoint,
+    data,
+  });
+};
+
+export const del = async <T>(endpoint: string): Promise<T> => {
+  const response = await apiClient.delete<T>(endpoint);
+  return response.data;
+};
+
+export default apiClient;
