+++ src/app/_types/image.types.ts
@@ -1,0 +1,7 @@
+// 이미지 타입 상수
+export const IMAGE_TYPES = [
+  'image/jpeg',
+  'image/png',
+  'image/jpg',
+  'image/webp',
+] as const;
