+++ src/app/_types/post.types.ts
@@ -14,7 +14,7 @@ export interface Post {
   postImageUrl: string | null;
   user: {
     name: string;
-    profileImgUrl: string | null;
+    profileImageUrl: string | null;
   };
   createdAt: string;
   updatedAt: string | null;
