+++ src/app/profile/_queries.ts
@@ -1,5 +1,5 @@
 import { queryOptions } from '@tanstack/react-query';
-import { getMyProfile } from '@/app/_apis/client';
+import { getMyProfile } from '@/app/_apis/user-api';
 
 export const profileQueries = {
   myProfile: () =>
